+++ ts/packages/memory/conversation/src/index.ts
@@ -11,3 +11,4 @@ export * from "./emailMessage.js";
 export * from "./emailImport.js";
 export * from "./docMemory.js";
 export * from "./docImport.js";
+export * from "./termParser.js";
